Update 2FA informaiton
Linked to the Zendesk section where relevant, as that's a more natural place for people to look for that information - we're most likely to come across this issue while on support. I deleted as much as I could here to avoid duplication, but so that it still makes sense for new people reading the team manual to understand our processes.