Replace the team dashboard page
With a Monitoring page. Since the move to remote working, there are no
team TVs, so frame this information in terms of monitoring in
general. I've set a short review time (1 month) as this seems likely
to change in the coming months.

I made a few changes when reading through this. The lines are wrapped
for readability, and I've replaced the [here] links, which aren't very
descriptive. I've also moved more links to use separate definitions,
as I think this also helps with readability.

The Trello, Google Meet and Google Calendar sections have gone, but
I'm not sure there was much information in there to keep.