Various small changes to enable local build/serve
Make the Ruby versions in the Dockerfile and .ruby-version
consistent. Correct the out-of-date alpine package name
for npm. Remove the unneeded 'npm install' invocation
which seems to have been a hangover from when the project
had a package.json (did we get the govuk-frontend as a
node pakage rather than a gem?).