Fix YAML parse errors caused by upgrade to Ruby 3.1.2
To overcome errors thrown by the latest version of the YAML
parser it has been necessary to overload a method in the
GovUK Tech Docs template/gem.

The team responsible have been contacted in order to see
whether we can get this changes in the gem. If so, we will
be able to revert this element of the commit but will have to
retain the other part, the quoted dates.